apps/ai-service/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
numpy>=1.26.0
face_recognition>=1.3.0
Pillow>=10.0.0
dlib==20.0.0
